+++ src/main/java/ladder/domain/randomGenerator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.randomGenerator;
 
 public interface NumberGenerator {
